refactor: move keybinding creation logic to a new module and update references. added comments to function files
