+++ src/extension.js
@@ -1,18 +1,18 @@
 const vscode = require('vscode');
-const createKeybinding = require('../Test_Keybinding_creation/createKeybinding');
+const createKeybinding = require('../Keybinding-Creation/createKeybinding');
 
-let dynamicKeybindingsEnabled = true;
-let activeProfile = 'profile1';
-let statusBarItem;
+let dynamicKeybindingsEnabled = false; // Define the initial state of dynamic keybindings
+let activeProfile = 'profile1'; // Define the initial profile
+let statusBarItem; // Status bar item to show the current state of dynamic keybindings
 
-function updateStatusBar() {
+function updateStatusBar() { // Update the status bar item with the current state of dynamic keybindings and active profile
     if (statusBarItem) {
         statusBarItem.text = `$(keyboard) ${dynamicKeybindingsEnabled ? 'Enabled' : 'Disabled'} - ${activeProfile}`;
         statusBarItem.show();
     }
 }
 
-function activate(context) {
+function activate(context) { // Define the status and create the bar item to show the current state of dynamic keybindings and active profile
     createKeybinding.activate(context);
     vscode.commands.executeCommand('setContext', 'dynamicKeybindingsEnabled', dynamicKeybindingsEnabled);
     vscode.commands.executeCommand('setContext', 'activeProfile', activeProfile);
